4x7-Segment-Display project
This was supposed to be a driver for a small board I made that had a 4 in 1 7 segment display and resistors. It was supposed to be able to pint out float numbers, but some half-way I realised I don't have the knowledge at this time to make everything work as intended, so I turned this into something I could practice, learn and have fun with while improving my STM programming capabilities. I was 17 at the time of writing this.